Use custom logo also in webmail
fix #3648
