+++ frontend/src/components/tools/CreationForm.vue
@@ -7,7 +7,11 @@
     >
       <ConfirmDialog ref="confirm" />
       <v-stepper-header class="align-center px-10">
-        <v-img :src="logoPath" max-width="190" class="hidden-sm-and-down" />
+        <v-img
+          :src="creationFormLogoPath"
+          max-width="190"
+          class="hidden-sm-and-down"
+        />
         <v-stepper-item
           v-for="(step, index) in steps"
           :key="step.name"
@@ -101,13 +105,13 @@
 <script setup lang="js">
 import ConfirmDialog from './ConfirmDialog'
 import CreationSummary from './CreationSummary.vue'
-import { computed, ref } from 'vue'
+import { ref } from 'vue'
 import { useGettext } from 'vue3-gettext'
 import { useDisplay } from 'vuetify'
-import { useGlobalConfig } from '@/main'
+import { useLogos } from '@/composables/logos'
 
 const { smAndDown } = useDisplay()
-const globalConfig = useGlobalConfig()
+const { creationFormLogoPath } = useLogos()
 
 const props = defineProps({
   title: { type: String, default: '' },
@@ -121,15 +125,6 @@ const props = defineProps({
 const currentStep = ref(1)
 const working = ref(false)
 const confirm = ref()
-
-const logoPath = computed(() => {
-  if (globalConfig.CREATION_FORM_LOGO_PATH) {
-    return globalConfig.CREATION_FORM_LOGO_PATH
-  }
-  const imgUrl = new URL('@/assets/Modoboa_RVB-BLEU-SANS.png', import.meta.url)
-    .href
-  return imgUrl
-})
 
 const emit = defineEmits(['close', 'create', 'validationError'])
